Major fixes for STORESTACK commands
Fixed a major bug causing many scripts using the STORESTACK command break the stack. Still seem to be stack issues but it's a lot better than it was.  Issues with multiple return nodes are known and being worked on. Also have a new idea for how to infer arg numbers, arg types, and return values.